modified:   src/main/java/com/conf/Config.java 	modified:   src/main/java/com/kafka/consumer/TheConsumer.java 	modified:   src/main/java/com/kafka/vo/KafkaMessage.java 	modified:   src/main/java/com/log/FileZIP.java 	modified:   src/main/java/com/log/LogFile.java 	modified:   src/main/java/com/mail/send/MailSend.java 	modified:   src/main/java/com/timer/StartTimer.java 	modified:   src/main/java/com/timer/Task.java

+++ MailNotify/src/main/java/com/log/FileZIP.java
@@ -8,15 +8,30 @@ import java.io.InputStream;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
+/**
+ * @author zuolang zip文件压缩
+ *
+ */
 public class FileZIP {
 
-	public void zip(String filepath, String zipFileName) { // 方法重载
+	/**
+	 * 生成压缩文件 
+	 * @param topic
+	 * @return 是否生成成功
+	 */
+	public boolean zip(String topic) { // 方法重载
+		System.out.println("日志文件：" + topic);
+		String zipName = topic+ ".zip";
+		System.out.println("zip文件名称：" + zipName);
 		InputStream input = null;
 		ZipOutputStream zipOut = null;
 		try {
 
-			File file = new File(filepath);
-			File zipFile = new File(zipFileName);
+			File file = new File(topic);
+			if(!file.exists()){
+				return false;
+			}
+			File zipFile = new File(zipName);
 			input = new FileInputStream(file);
 			zipOut = new ZipOutputStream(new FileOutputStream(zipFile));
 			zipOut.putNextEntry(new ZipEntry(file.getName()));
@@ -36,5 +51,6 @@ public class FileZIP {
 				e.printStackTrace();
 			}
 		}
+		return true;
 	}
 }
